backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
supabase
httpx
pulp
python-dotenv
